+++ test/day8/re_frame/async_flow_fx_test.cljs
@@ -37,21 +37,21 @@
 
 (deftest test-massage-rules
   (is (= (core/massage-rules :my-id [{:when :seen? :events :1 :dispatch [:2]}])
-         (list {:id 0 :when core/seen-all-of? :events #{:1} :dispatch (list [:2])})))
+         (list {:id 0 :when core/seen-all-of? :events #{:1} :dispatch-n (list [:2])})))
 
   (is (= (core/massage-rules :my-id [{:when :seen-both? :events [:1 :2] :halt? true}])
-         (list {:id 0 :when core/seen-all-of? :events #{:1 :2} :dispatch (list [:my-id :halt-flow])})))
+         (list {:id 0 :when core/seen-all-of? :events #{:1 :2} :dispatch-n (list [:my-id :halt-flow])})))
 
   (is (= (core/massage-rules :my-id [{:when :seen-any-of? :events #{:1 :2} :dispatch [:2] :halt? true}])
-         (list {:id 0 :when core/seen-any-of? :events #{:1 :2} :dispatch (list [:2] [:my-id :halt-flow])}))))
+         (list {:id 0 :when core/seen-any-of? :events #{:1 :2} :dispatch-n (list [:2] [:my-id :halt-flow])}))))
 
 
-(deftest test-steup
-  (let [flow {:first-dispatch [:1]
-              :rules [
-                      {:when :seen? :events :1 :dispatch [:2]}
-                      {:when :seen? :events :3 :halt? true}]}
-        handler-fn   (core/make-flow-event-handler flow)]
+(deftest test-setup
+  (let [flow       {:first-dispatch [:1]
+                    :rules          [
+                                     {:when :seen? :events :1 :dispatch [:2]}
+                                     {:when :seen? :events :3 :halt? true}]}
+        handler-fn (core/make-flow-event-handler flow)]
     (is (= (handler-fn {:db {}} [:dummy-id :setup])
            {:db             {}
             :dispatch       [:1]
@@ -89,7 +89,7 @@
                        :rules-fired #{}}}}
              [:test-id [:1]])
            {:db {:p {:seen-events #{:1} :rules-fired #{0}}}
-            :dispatch (list [:2])}))
+            :dispatch-n (list [:2])}))
 
     ;; new event should cause a dispatch
     (is (= (handler-fn
@@ -97,7 +97,7 @@
                        :rules-fired #{0}}}}
              [:test-id [:3]])
            {:db {:p {:seen-events #{:1 :3} :rules-fired #{0 1}}}
-            :dispatch (list [:test-id :halt-flow])}))
+            :dispatch-n (list [:test-id :halt-flow])}))
 
     ;; make sure :seen-any-of? works
     (is (= (handler-fn
@@ -105,7 +105,7 @@
                        :rules-fired #{}}}}
              [:test-id [:4]])
            {:db {:p {:seen-events #{:4} :rules-fired #{2}}}
-            :dispatch (list [:6])}))))
+            :dispatch-n (list [:6])}))))
 
 
 (deftest test-halt1